requirements.txt
numpy==1.23.3
Pillow==9.3.0
requests==2.28.1
streamlit_lottie==0.0.3
